First draft of an update guide
This isn't quite finished yet, but it got to a place where I want to
take a break from it, and there's enough there to give feedback on the
existing content. I figure for new guides we can start the discussion
here and then move to the website once we're good with the content. It's
easier to review without markup.